benchmarks: add benchmarks for TCE
