+++ etc/make-format.sh
@@ -13,5 +13,7 @@ find tests/*.hpp -type f ! -iname "catch.hpp" -exec clang-format -i --verbose {}
 find tests/*.cpp -type f -exec clang-format -i --verbose {} \;
 find benchmarks/*.hpp -type f ! -iname "catch.hpp" -exec clang-format -i --verbose {} \;
 find benchmarks/*.cpp -type f -exec clang-format -i --verbose {} \;
+find benchmarks/examples/*.hpp -type f -exec clang-format -i --verbose {} \;
+find benchmarks/examples/*.cpp -type f -exec clang-format -i --verbose {} \;
 
 exit 0
